Fix overflow when passing thru hole in wall in sid.wad's MAP01
Reverting part of commit fe81b0e88af7aad78299c232042f2737f63da522.
Shouldn't have changed this in the first place.

+++ src/r_segs.c
@@ -316,7 +316,7 @@ void R_RenderMaskedSegRange(drawseg_t *ds, int x1, int x2)
                     || t > (int64_t)SCREENHEIGHT << FRACBITS * 2)
                     continue;                    // skip if the texture is out of screen's range
 
-                sprtopscreen = (int)(t >> FRACBITS);
+                sprtopscreen = (int64_t)(t >> FRACBITS);
             }
 
             dc_iscale = 0xFFFFFFFFu / (unsigned int)spryscale;
